generic client factory
